+++ rust/src/benchmark/bencher.rs
@@ -2,16 +2,24 @@ use std::{path::Path, sync::Arc};
 
 use itertools::Itertools;
 use libdlock::{
+    dlock::rcl::{rcllock::RclLock, rclserver::RclServer},
     dlock::{BenchmarkType, DLockType},
     parker::{block_parker::BlockParker, spin_parker::SpinParker, Parker},
-    dlock::rcl::{rcllock::RclLock, rclserver::RclServer},
 };
 
 use strum::IntoEnumIterator;
 
 use crate::{
     benchmark::{
-        dlock2::benchmark_dlock2, non_cs_counter::counter_one_three_non_cs_one, one_three_ratio_counter::counter_one_three_benchmark, proposion_counter::counter_proportional, response_time_single_addition::benchmark_response_time_single_addition, response_time_variable::benchmark_response_time_one_three_ratio, subversion_job::counter_subversion_benchmark
+        dlock::{
+            non_cs_counter::counter_one_three_non_cs_one,
+            one_three_ratio_counter::counter_one_three_benchmark,
+            proposion_counter::counter_proportional,
+            response_time_single_addition::benchmark_response_time_single_addition,
+            response_time_variable::benchmark_response_time_one_three_ratio,
+            subversion_job::counter_subversion_benchmark,
+        },
+        dlock2::benchmark_dlock2,
     },
     command_parser::{experiment::Experiment, lock_target::*},
 };
