+++ rust/src/benchmark/dlock/proposion_counter.rs
@@ -26,7 +26,7 @@ use libdlock::{
 };
 use quanta::Clock;
 
-use super::bencher::LockBenchInfo;
+use crate::benchmark::bencher::LockBenchInfo;
 
 thread_local! {
     static WRITER: OnceCell<
